requirements.txt
GitPython==3.1.40
javalang==0.13.0
lxml==5.1.0
